x and y of Position now double

+++ src/main/java/it/unibo/michelito/util/position/Position.java
@@ -1,4 +1,4 @@
 package it.unibo.michelito.util.position;
 
-public record Position(int x, int y) {
+public record Position(double x, double y) {
 }
